+++ src/add-ons/super-forms-signature/assets/js/frontend/jquery.signature.js
@@ -135,6 +135,14 @@ $(selector).signature({color: 'blue', guideline: true}) */
 			}
 		},
 
+		// Added for Super Forms Conditional Logic, so that the canvas is correctly resized after the column is visible.
+		resize: function() {
+			console.log('resizing');
+			var parent = $(this.canvas).parents('.super-signature-canvas:eq(0)');
+			if($(this.canvas).width()!=parent.width()) this.canvas.setAttribute('width', parent.width());
+			if($(this.canvas).height()!=parent.height()) this.canvas.setAttribute('height', parent.height());
+		},
+
 		/** Synchronise changes and trigger a change event.
 			@memberof Signature
 			@private